入力をdouble型にする

+++ sample.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 using namespace std;
  int main(){
-  int A, B;
+  double A, B;
   string op;
   //「3 * 4」のように，（整数）[スペース]（演算子）[スペース]（整数）の形で入力
   cin >> A >> op >> B;
@@ -16,7 +16,7 @@ using namespace std;
     if(B==0){
       cout<<"Error"<<endl;
     }else{
-      cout<<A / B<<endl; 
+      cout<<(double) A / B<<endl; 
   }}else{
       cout<<"Error"<<endl;
   }
